+++ src/core_plugins/kibana/public/dashboard/index.html
