+++ src/core_plugins/kibana/public/dashboard/styles/index.less
